requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
gradio>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
peft>=0.7.0
diffusers>=0.24.0
opencv-python>=4.8.0
scipy>=1.11.0
scikit-image>=0.22.0
